requirements.txt
sqlalchemy~=1.3.19
psycopg2
asyncpg
alembic==1.4.2
pytelegrambotapi==3.7.2
textract==1.6.3
requests~=2.24.0
Pillow~=7.2.0